test and documentation updates to ActorId
- moved tests from DocTests into their own file in the main test suite
  and broke them up by functionality tested
- added documentation to ActorId and its new initializer options
- updated warning content for the error condition of using the same
  actorId concurrently
