+++ Sources/Automerge/ActorId.swift
@@ -3,8 +3,14 @@ import Foundation
 
 /// The identifier for collaborators contributing to an Automerge document.
 ///
-/// Each separate instance of an Automerge document should have it's own, unique, `ActorId`.
-/// If you create your own `ActorId`, no concurrent changes should ever be made with the same `ActorId`.
+/// Each instance of an Automerge document should have its own, unique, `ActorId`.
+/// A new, or forked, Automerge document automatically creates its own ActorId.
+///
+/// > Warning: It is an error in Automerge to make two concurrent changes to an Automerge document with the same
+/// ActorId.
+/// When you create your own ActorIds, you are responsible to enforce that two concurrent changes on a document does not
+/// happen.
+/// Because of this, Automerge recommends that you use the default ActorId that it provides.
 public struct ActorId: Equatable, Hashable, Sendable {
     var data: Data
 
@@ -12,17 +18,26 @@ public struct ActorId: Equatable, Hashable, Sendable {
         data = Data(ffi)
     }
 
-    // Creates a new, random actor.
+    /// Creates a random Actor Id.
     public init() {
         self.init(uuid: UUID())
     }
 
-    // Creates an actor from a UUID.
+    /// Creates an Actor Id from the contents of a UUID you provide.
     public init(uuid: UUID) {
-        data = withUnsafeBytes(of: uuid.uuid, { Data($0) })
+        data = withUnsafeBytes(of: uuid.uuid) { Data($0) }
     }
 
-    // Creates an actor from data.
+    /// Creates an Actor Id from the data you provide.
+    /// - Parameter data: A maximum of 128 bytes that represents the Actor Id.
+    ///
+    /// ActorId imposes a limit of 128 bytes for the data that represents the actor Id.
+    ///
+    /// > Warning: It is an error in Automerge to make two concurrent changes to an Automerge document with the same
+    /// ActorId.
+    /// When you create your own ActorIds, you are responsible to enforce that two concurrent changes on a document does
+    /// not happen.
+    /// Because of this, Automerge recommends that you use the default ActorId that it provides.
     public init?(data: Data) {
         guard data.count <= 128 else {
             return nil
@@ -32,6 +47,7 @@ public struct ActorId: Equatable, Hashable, Sendable {
 }
 
 extension ActorId: CustomStringConvertible {
+    /// A hex-encoded string that represents the bytes of the Actor Id.
     public var description: String {
         data.map { String(format: "%02hhX", $0) }.joined()
     }
