--- Tests/AutomergeTests/ActorTests.swift
@@ -1,10 +1,0 @@
-@testable import Automerge
-import XCTest
-
-class ActorTests: XCTestCase {
-    func testDefaultActorSize() {
-        let doc1 = Document()
-        XCTAssertEqual(doc1.actor.bytes.count, 16)
-    }
-}
-
